task: removes warnings

+++ tasks/Task.cpp
@@ -162,7 +162,7 @@ void Task::updateHook()
 		if (joints_resurrection[i]<3)
                 {
 			//! In case of reseting a node stop the motion of the rest of the motor
-                        for (register size_t j=0;j<numMotors;j++)
+                        for (register size_t j=0;j<static_cast<size_t>(numMotors);j++)
                         {
                                 if (canParameters.Active[j])
                                 {
@@ -178,7 +178,7 @@ void Task::updateHook()
 			if (canParameters.Active[i]==ACTIVE)
                         {	
 			    //! In case of inactivating a node stop the motion of the rest of the motor
-                            for (register size_t j=0;j<numMotors;j++)
+                            for (register size_t j=0;j<static_cast<size_t>(numMotors);j++)
                             {
                                 if (canParameters.Active[j])
                                 {
